+++ src/FSharp.Configuration.Runtime/AssemblyInfo.fs
@@ -2,7 +2,7 @@
 namespace System
 open System.Reflection
 
-[<assembly: AssemblyTitleAttribute("FSharp.Configuration")>]
+[<assembly: AssemblyTitleAttribute("FSharp.Configuration.Runtime")>]
 [<assembly: AssemblyProductAttribute("FSharp.Configuration")>]
 [<assembly: AssemblyDescriptionAttribute("The FSharp.Configuration project contains type providers for the configuration of .NET projects.")>]
 [<assembly: AssemblyVersionAttribute("2.0.0")>]
@@ -10,7 +10,7 @@ open System.Reflection
 do ()
 
 module internal AssemblyVersionInformation =
-    let [<Literal>] AssemblyTitle = "FSharp.Configuration"
+    let [<Literal>] AssemblyTitle = "FSharp.Configuration.Runtime"
     let [<Literal>] AssemblyProduct = "FSharp.Configuration"
     let [<Literal>] AssemblyDescription = "The FSharp.Configuration project contains type providers for the configuration of .NET projects."
     let [<Literal>] AssemblyVersion = "2.0.0"