+++ src/FSharp.Configuration.Runtime/FSharp.Configuration.Runtime.fs
@@ -1,0 +1,5 @@
+namespace FSharp.Configuration.Yaml
+
+// Put the TypeProviderAssemblyAttribute in the runtime DLL, pointing to the design-time DLL
+[<assembly:CompilerServices.TypeProviderAssembly("FSharp.Configuration.DesignTime.dll")>]
+do ()